quranapp_backend/requirements.txt
uvicorn~=0.23.2
fastapi~=0.103.2
SQLAlchemy~=2.0.21
psycopg2~=2.9.9
alembic~=1.12.0
azure-storage-blob~=12.18.2
pydantic~=2.3.0
python-multipart~=0.0.6
pyjwt~=2.8.0
